sys_pre_expand: Don't duplicate options given in the source code
Any compiler options given with a -compile() attribute in source file
would be included both at the beginning and the end of the option list
for the compiler. Including the options twice is harmless during
compilation, but since the options will also be available in
Mod:module_info(compile), including them twice will waste memory.

Include the options from the source first in the list so that they
override options given on the command line.