erts/configure.in: Remove test for reversed setvbuf() arguments
The test is not needed, because no code tests whether SETVBUF_REVERSED
is defined.

Furthermore, AC_FUNC_SETVBUF_REVERSED is now documented like this:

Do nothing. Formerly, this macro checked whether setvbuf takes the
buffering type as its second argument and the buffer pointer as the
third, instead of the other way around, and defined
SETVBUF_REVERSED. However, the last systems to have the problem were
those based on SVR2, which became obsolete in 1987, and the macro is
no longer needed.