compiler: Don't include source code options in module_info(compile)
As far as I know, the purpose of the compiler options included in
Mod:module_info(compile) has never been documented.

An educated guess is that they are there in case you want to
re-compile the module with the same options, and also to aid in
debugging when you need to know how a module was compiled.

In neither case is there any need to include options given in the
source itself in options included in Mod:module_info(compile).
Including those options will only waste memory.

Therefore, only include in the BEAM file the options that were
given to compile:file/2.