tar_SUITE: Work around limited pathname length on Windows
Ever since the test suites were converted to the common_test
convention, the tests for long names started to fail on Windows. The
reason is that those test cases depend on the current working
directory, and that the current working directory in common_test test
cases is usually set to a directory with a longer pathname than in
test_server test suites.

Since neither Erlang/OTP nor common_test provides a portable way
to create a temporary directory with a short name, we will have to
abuse priv_dir by going upwards from it. For further information,
see the comment in run_in_short_tempdir/2.