beam_jump: Fix broken optimization
A test instruction with the same target label as jump immediately
followed it was supposed to be removed, but it was kept anyway.

Fix that optimization, but also make sure that the test instruction
is kept if the test instruction may have side effects (such as
a bit syntax matching instruction).

While at it, make the code cleaner by breaking it up into two clauses
and don't remove the jump instruction if it is redudant (removal of
redundant jump instructions is already handled in another place).