compiler: Eliminate use of deprecated erlang:hash/2
Now that beam_asm computes the Index and Uniq values for funs, there
is no need to compute those values in the sys_pre_expand and v3_kernel
modules, thus eliminating the calls to the deprecated erlang:hash/2
function.

It would be tempting to stop generating the name for the fun in
sys_pre_expand so that we did not have to add the Info field to a tuple.

But:

* The debugger depends on the name being there. (Simple solution:
  Let the debugger generate the name itself.)

* When a fun has been inlined into another function, the fun name in
  'id' annotation will be used to notice the inlining and change the
  final clause of the top-level case from generating a 'function_clause'
  exception to a case_clause exception. (Possible workaround: Have the
  inliner set an inlined attribute on functions that have been inlined, or
  have the inliner rewrite 'function_clause' exceptions itself.)