beam_load.c: Remove useless call to next_heap_size()
The idea was probably to cause less fragmentation. Even if that would
be true, it is irrelevant because the short-lived allocator that is
used does not have any problems with fragmentation.

In CodeNeed() we will simply double the size of the area used for code
instead of using the next heap size.