erl_process.c: Remove the redundant variable 'processes_busy'
There is a static variable called 'processes_busy' in erl_process.c,
which will be incremented but never used.

To confuse things, there is also a global variable with the same
name, but it it is only defined and used if BM_COUNTERS is defined.
In erl_process.c, the global version of 'processes_busy' will be
seen and incremented if BM_COUNTERS is defined.

Remove the static version of 'processes_busy' in erl_process.c,
but keep the global version if BM_COUNTERS is defined.

Noticed-by: Jovi Zhang