Fix dialyzer warnings in supervisor
Dialyzer complained over a mismatch between the callback spec of
Mod:code_change in gen_server and the spec of supervisor:code_change
(which is the implementation of a gen_server Mod:code_change).

This commit changes the callback spec to allow {error,Reason} as
return value. Also, release_handler is updated to handle this return
value.