Keep unpack_release backwards compatible with old systools:make_tar
systools:make_tar has recently been updated to store the .rel file
in both releases and releases/<vsn> directories. Ths same commit
removed copying of this file from releases to releases/<vsn> in
release_handler:unpack_release. This commit adde this copy again to
make sure the file is not lost if the tar file was created with an
older systools.