Add the beam_except pass to optimize exceptions
In order to save space, rewrite suitable calls to erlang:error/{1,2}
to special BEAM instructions.

This code is probably longer than the code taken out of v3_life and
v3_codegen in the previous commit, but it is much easier to
understand and maintain since the BEAM assembler format is better
understood than the v3_life format.