compilation_SUITE: Prevent cover from being run on slave nodes
In the self compilation test cases, the compiler compiles itself
and runs the newly compiled version on a slave node. Having the
cover server starting on the slave node defeats the purpose of
the test, since it will load the SAME cover-compiled code on the
slave node. (It will also be slower, but will not improve coverage
since it compiles the same source files again.)

Use a shielded node to prevent the cover server from getting
started on the slave node.