bif_SUITE: Add test case for specs and BIF stubs
Rename the existing types/1 test case to erl_bif_types/1 to make
it clearer what it does. Also no longer test for missing type
information for BIFs (since it has become optional), but only
check that the information provided seems to be consistent.

Introduce the specs/1 test case to ensure that all BIFs have
specs.

Also introduce improper_bif_stubs/1 to check for proper stubs.
Since the BEAM loader will now silently allow stubs for BIFs, we
want to be particular about exactly what a stub look like, so that
an Erlang function is not unintentionally overridden by a BIF.