Set `font-family: Courier, monospace' in OTP doc CSS
This should be much better for everybody, I left Courier as the primary
original font and also added monospace as secondary for people like me
which in Linux haven't it installed.

It should be more pleasant to read typespecs and code examples. Also
adds minor cosmetic changes to the CSS.