erl_lint: Consistently reject local calls from guards
If a guard test (such as is_list/1) has a local definition in a
module (or is imported), erl_lint will reject a call to it from
a guard if the call is not at the top-level:

foo(L) when is_list(L) =:= true ->   %% Will be rejected.
   ok.

is_list(_) -> ok.

But if the call is at the top-level, it will be accepted (and
cause a crash in a later compiler pass):

foo(L) when is_list(L) ->   %% Will be accepted by erl_lint
   ok.

is_list(_) -> ok.

This inconsistency was an oversight introduced when it became
possible to override BIFs with local definitions.