erts: Correct incorrect assertion in erl_alloc_util.c
The incorrect assertion would fire when '+Muramv true' has been
given, i.e. when a reallocation always forces a move.

Change the assertion to not call the lock checker if the mutex is
not used (because the allocator will only be accessed from a
specific scheduler and thus no lock is needed).