Replace autoimport_SUITE with bif_SUITE:auto_imports/1
erts/test/autoimport_SUITE tested that auto-import information in
erl_internal:bif/2 was consistent with the documentation. It did
it by scanning erlang.xml.

Since the documentation is now based on the specs in erlang.erl,
we should now test consistency of the specs and erl_internal:bif/2.

Since anyone that adds a new BIF runs the emulator test suite, it
makes sense to do this test in bif_SUITE in the emulator test suite.