Clean up lib/Makefile
We only support building epmd, erl_interface, and ic for VxWorks;
therefore there is no need to have a conditional section for VxWorks.

We use to build different applications on Windows and Unix, but
that is no longer the case; therefore one list of applications is
sufficient.

Unfortunately, there are build order dependencies between the
applications, so we must not sort the list.