[observer] Windows double buffer fixes
DC's and GC's is not double buffered by default on windows,
and there is a separate erase event which causes awful flickering
when constant updating a window.

Hack around wx (to be able to use wxBufferPaintDC),
to avoid flickering on windows.

This works on windows because there (and only there)
wxGC:create/1 also takes a memoryDC as argument.