erts: Refactor code loading with renaming
Rename lock_code_ix as seize_code_write_permission. Don't want to call
it a "lock" as it can be held between schedulings and different threads
and is not managed by lock checker.

Rename "activate" staging as "commit" staging. Why not be consistent
and use git terminology all the way.