[reltool] Unset ERL_FLAGS when running escript from reltool_server_SUITE
OTP-9794

Some escript tests in reltool_server_SUITE failed if ERL_FLAGS was set
to enable smp. The reason is that the test expected smp_support to be
false, since the argument "-smp disable" is stated in the %%!-line of
the escript. However, ERL_FLAGS will override the arguments in the
escript so the result could not be truested.